Extract breadcrumb component to a separate file
