+++ ui/src/components/breadcrumbs.module.scss
@@ -1,0 +1,26 @@
+@import 'design-system/variables/colors.scss';
+@import 'design-system/variables/typography.scss';
+
+.breadcrumbs {
+  display: flex;
+  align-items: center;
+
+  :not(:last-child) {
+    margin-right: 8px;
+  }
+}
+
+.breadcrumb {
+  display: block;
+  @include paragraph-x-small();
+  font-weight: 600;
+  color: $color-neutral-800;
+
+  &.link {
+    color: $color-primary-1-600;
+
+    &:focus-visible {
+      box-shadow: 0 0 0 2px $color-generic-black;
+    }
+  }
+}
